requirements.txt
langchain>=0.3,<0.4
langchain-openai>=0.3,<0.4
langchain-mcp-adapters>=0.1
python-dotenv>=1.0,<2.0
langgraph>=0.4,<0.5
aiohttp>=3.9,<4.0
tenacity>=8.0,<10.0
pytest>=8.0
pytest-asyncio>=0.23
aioresponses>=0.7
